# Improve reading with `use_stream = TRUE`