--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -107,7 +107,7 @@ Suggests:
     tidyselect (>= 1.0.0),
     tmap (>= 2.0),
     vctrs,
-    wk
+    wk (>= 0.9.0)
 LinkingTo: 
     Rcpp
 VignetteBuilder: