--- a/R/read.R
+++ b/R/read.R
@@ -239,9 +239,12 @@ process_cpl_read_ogr_stream = function(x, default_crs, num_features, fid_column_
 		st_set_crs(x, crs)	
 	})
 
-	# Prefer "geometry" as the geometry column name
+	# Prefer "geometry" as the geometry column name instead of "wkb_geometry"
 	if (any(is_geometry_column) && !("geometry" %in% names(df))) {
-		names(df)[which(is_geometry_column)[1]] = "geometry"
+		geometry_column_name <- names(df)[which(is_geometry_column)[1]]
+		if (identical(geometry_column_name, "wkb_geometry")) {
+			names(df)[which(is_geometry_column)[1]] = "geometry"
+		}
 	}
 	
 	# Rename OGC_FID to fid_column_name and move to end