--- a/R/read.R
+++ b/R/read.R
@@ -218,46 +218,46 @@ default_st_read_use_stream = function() {
 	)
 }
 
-process_cpl_read_ogr_stream = function(x, default_crs, num_features, fid_column_name,
+process_cpl_read_ogr_stream = function(x, geom_column_info, num_features, fid_column_name,
                                        crs = NULL, promote_to_multi = TRUE, ...) {
-    is_geometry_column = vapply(
+	is_geometry_column = vapply(
 		x$get_schema()$children,
 		function(s) identical(s$metadata[["ARROW:extension:name"]], "ogc.wkb"),
 		logical(1)
 	)
-
-    crs = if (is.null(crs)) st_crs(default_crs) else st_crs(crs)
-	if (num_features == -1) {
+    
+    geom_column_info$index = which(is_geometry_column)
+    
+    if (num_features == -1) {
 		num_features = NULL
 	}
+    
+    # Suppress warnings about extension type conversion (since we want the
+    # default behaviour of converting the storage type)
 	df = suppressWarnings(nanoarrow::convert_array_stream(x, size = num_features))
-
-	df[is_geometry_column] = lapply(df[is_geometry_column], function(x) {
-		attributes(x) <- NULL
+	
+	for (i in seq_len(nrow(geom_column_info))) {
+		crs = if (is.null(crs)) st_crs(geom_column_info$crs[[i]]) else st_crs(crs)
+		name = geom_column_info$name[[i]]
+		index = geom_column_info$index[[i]]
 		
-		x <- wk::wk_handle(wk::new_wk_wkb(x), wk::sfc_writer(promote_multi = promote_to_multi))
-		st_set_crs(x, crs)	
-	})
-
-	# Prefer "geometry" as the geometry column name instead of "wkb_geometry",
-	# which is what happens when the geometry column doesn't have a name
-	if (any(is_geometry_column) && !("geometry" %in% names(df))) {
-		geometry_column_name <- names(df)[which(is_geometry_column)[1]]
-		if (identical(geometry_column_name, "wkb_geometry")) {
-			names(df)[which(is_geometry_column)[1]] = "geometry"
-		}
+		column_wkb = df[[index]]
+		attributes(column_wkb) = NULL
+		column_sfc = wk::wk_handle(
+			wk::new_wk_wkb(column_wkb),
+			wk::sfc_writer(promote_multi = promote_to_multi)
+		)
+		
+		df[[index]] = st_set_crs(column_sfc, crs)
+		names(df)[index] = name
 	}
 	
 	# Rename OGC_FID to fid_column_name and move to end
 	if (length(fid_column_name) == 1 && "OGC_FID" %in% names(df)) {
-		df <- df[c(setdiff(names(df), "OGC_FID"), "OGC_FID")]
+		df = df[c(setdiff(names(df), "OGC_FID"), "OGC_FID")]
 		names(df)[names(df) == "OGC_FID"] = fid_column_name
 	}
 	
-	# Move geometry to the end
-#	if ("geometry" %in% names(df)) {
-#		df <- df[c(setdiff(names(df), "geometry"), "geometry")]
-#	}
 	gc1 = which(is_geometry_column)[1]
 	df = df[c(setdiff(seq_along(df), gc1), gc1)]
 
@@ -301,7 +301,8 @@ st_read.character = function(dsn, layer, ..., query = NA, options = NULL, quiet 
 		stream = nanoarrow::nanoarrow_allocate_array_stream()
 		info = CPL_read_gdal_stream(stream, dsn, layer, query, as.character(options), quiet,
 		    drivers, wkt_filter, dsn_exists, dsn_isdb, fid_column_name, getOption("width"))
-		process_cpl_read_ogr_stream(stream, default_crs = info[[1]], num_features = info[[2]],
+		geom_column_info = data.frame(name = info[[1]], crs = info[[2]], stringsAsFactors = FALSE)
+		process_cpl_read_ogr_stream(stream, geom_column_info, num_features = info[[3]],
 			fid_column_name = fid_column_name, stringsAsFactors = stringsAsFactors, quiet = quiet,
 			promote_to_multi = promote_to_multi, ...)
 	} else {